requirements.txt
streamlit>=1.35.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0